Refactor file structure to have performance ratings code in separate folder
